fix: Tests not running
Issue was that the rocket() method was async. This meant that
lazy_static could not be evaluated correctly. The solution to this
problem was to switch the Testing Client mutex from tokio to std.

The RWLock on settings was also tokio, but this could not be switched to
std because the endpoints that use SETTINGS are async. This meant I had
to use block_on to force the async call of settings in rocket() to be
sync